+++ gatsby-browser.js
@@ -1,19 +1,15 @@
 import * as React from "react";
-import { navigate } from "gatsby";
-import { AuthConfig } from "react-use-auth";
-import { Auth0 } from "react-use-auth/auth0";
 import "./src/styles/global.css";
+import { Auth0Provider } from "@auth0/auth0-react";
 
 export const wrapRootElement = ({ element }) => (
 	<>
-		<AuthConfig
-			navigate={navigate}
-			authProvider={Auth0}
-			params={{
-				domain: process.env.AUTH0_DOMAIN,
-				clientID: process.env.AUTH0_CLIENT_ID,
-			}}
-		/>
-		{element}
+		<Auth0Provider
+			domain={process.env.GATSBY_AUTH0_DOMAIN}
+			clientId={process.env.GATSBY_AUTH0_CLIENT_ID}
+			redirectUri={window.location.origin}
+		>
+			{element}
+		</Auth0Provider>
 	</>
 );
