+++ package.json
@@ -15,12 +15,14 @@
     "clean": "gatsby clean"
   },
   "dependencies": {
+    "@auth0/auth0-react": "^1.6.0",
     "auth0-js": "^9.16.2",
     "axios": "^0.21.1",
+    "express-jwt": "^6.1.0",
     "gatsby": "^3.10.2",
+    "jwks-rsa": "^2.0.4",
     "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-use-auth": "^2.1.0-14"
+    "react-dom": "^17.0.1"
   },
   "devDependencies": {
     "@netlify/plugin-gatsby": "^1.0.3",
